requirements.txt
pydantic>=2.7
python-dateutil
rapidfuzz>=3.9
pandas>=2.2
PyYAML>=6.0
requests
beautifulsoup4
lxml
imapclient>=2.3
mail-parser>=3.15
